requirements.txt
google_api_python_client==2.130.0
huggingface_hub==0.23.1
numpy==1.26.4
openai==0.28.1
Pillow==10.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
streamlit==1.35.0
streamlit_option_menu==0.3.12
transformers==4.41.1
psycopg2-binary == 2.9.9